Fix RGB LEDs colour order
The current colour arrangement used for DotStar has Green and Red swapped.
At least on my unit this is what I experience.
This PR fixes the order to `DOTSTAR_BGR`, but it might be useful to check if different batches of APA102c ordered for production might have different colour order