Use major version ref of `carlosperate/download-file-action` (#44)
The `carlosperate/download-file-action` action is used by the "Sync Labels" GitHub Actions workflow as a convenient way
to download external resources.

A major version ref has been added to that repository, meaning it is no longer necessary to do a full pin of the action
version in use as before.

Use of the major version ref will cause the workflow to use a stable version of the action, while also benefiting from
ongoing development to the action up until such time as a new major release of an action is made, at which time we would
need to evaluate whether any changes to the workflow are required by the breaking change that triggered the major
release before updating the major ref (e.g., uses: `carlosperate/download-file-action@v2`). I think this strikes the
right balance between stability and maintainability for the workflow.